v1.3
Major UI changes compared to 1.2, streamlining of information display & output. This is what I would consider the first "release-ready" version.

Planned features of v1.4:
Pressing the "enter" or "go" key on the android number pad will cause the focus to switch to the next lowest cell. When at the bottom cell, the "enter" or "go" key will trigger the "compute" button.